pyproject.toml
[tool.poetry]
name = "scrapy-first"
version = "0.1.0"
description = ""
authors = ["Yuriy Kormin <yuriy.kormin@gmail.com>"]
readme = "README.md"
packages = [{include = "chocolatescraper"}]

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.8.0"
ipython = "^8.12.0"
scrapy-playwright = "^0.0.26"
playwright = "^1.32.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
django = "^4.2"
django-bootstrap4 = "^22.3"
django-simple-history = "^3.3.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
